requirements.txt
# Web Framework
flask
flask-socketio
gevent
gevent-websocket

# Authentication
bcrypt

# SSH Server
asyncssh

# System Monitoring
psutil

# Windows Terminal Support
pywinpty; sys_platform == 'win32'
